requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
minio==7.2.0
celery==5.3.4
python-socketio[client]==5.10.0
prometheus-client==0.19.0
slowapi==0.1.9
structlog==23.2.0
pytest==7.4.3
python-multipart==0.0.6
amqp==5.2.0
kombu==5.3.4
billiard==4.2.0
vine==5.1.0
httpx==0.25.2
email-validator==2.1.0